# apps/server/public/swagger.yml
info:
  title: rs.school API
  version: 1.0.0
  description: ''
basePath: /api
swagger: '2.0'
paths:
  '/course/{courseId}/mentor/students':
    get:
      description: Returns mentors students results
      parameters:
        - name: courseId
          in: path
          description: Course Id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        '200':
          description: User object
  '/course/{courseId}/students':
    get:
      description: Returns course students
      security:
        - cookieAuth: []
      parameters:
        - name: courseId
          in: path
          description: Course Id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        '200':
          description: List of students
    post:
      description: Add/Update course students
      security:
        - cookieAuth: []
      parameters:
        - name: courseId
          in: path
          description: Course Id
          required: true
          type: integer
        - name: students
          in: body
          schema:
            type: array
            items:
              type: object
              properties:
                githubId:
                  type: string
                isExpelled:
                  type: boolean
                expellingReason:
                  type: boolean
                readyFullTime:
                  type: boolean
      produces:
        - application/json
      responses:
        '200':
          description: List of students
  '/course/{courseId}/externalAccounts':
    get:
      description: Returns external accounts of students
      security:
        - cookieAuth: []
      parameters:
        - name: courseId
          in: path
          description: Course Id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        '200':
          description: List of students with external accounts
  '/course/{courseId}/mentors':
    get:
      description: Saves course mentors
      security:
        - cookieAuth: []
      parameters:
        - name: courseId
          in: path
          description: Course Id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        '200':
          description: List of mentors
    post:
      description: Returns course mentors
      security:
        - cookieAuth: []
      parameters:
        - name: courseId
          in: path
          description: Course Id
          required: true
          type: integer
        - name: mentors
          in: body
          schema:
            type: array
            items:
              type: object
              properties:
                githubId:
                  type: string
                maxStudentsLimit:
                  type: integer
      produces:
        - application/json
      responses:
        '200':
          description: List of mentors
  '/course/{courseId}/pairs':
    post:
      description: Assign student to mentor
      security:
        - cookieAuth: []
      parameters:
        - name: courseId
          in: path
          description: Course Id
          required: true
          type: integer
        - name: pairs
          in: body
          schema:
            type: array
            items:
              type: object
              properties:
                studentGithubId:
                  type: string
                mentorGithubId:
                  type: string
      produces:
        - application/json
      responses:
        '200':
          description: List of pairs
  '/course/{courseId}/tasks':
    get:
      description: Returns course tasks
      parameters:
        - name: courseId
          in: path
          required: true
          type: integer
      produces:
        - application/json
      responses:
        '200':
          description: List of tasks object
  '/course/{courseId}/task':
    post:
      description: Assign task to course/stage
      parameters:
        - name: courseId
          in: path
          required: true
          type: integer
      produces:
        - application/json
      responses:
        '200':
          description: Result
  '/course/{courseId}/task/{courseTaskId}':
    put:
      description: Update course task
      parameters:
        - name: courseId
          in: path
          required: true
          type: integer
      produces:
        - application/json
      responses:
        '200':
          description: Result
    delete:
      description: Delete course task
      parameters:
        - name: courseId
          in: path
          required: true
          type: integer
      produces:
        - application/json
      responses:
        '200':
          description: Result
  '/course/{courseId}/task/{courseTaskId}/shuffle':
    post:
      description: Assign course task to checker
      parameters:
        - name: courseId
          in: path
          required: true
          type: integer
        - name: courseTaskId
          in: path
          required: true
          type: integer
      produces:
        - application/json
      responses:
        '200':
          description: Result
  '/course/{courseId}/scores':
    post:
      description: Save course task score
      security:
        - cookieAuth: []
      parameters:
        - name: courseId
          in: path
          required: true
          type: integer
        - name: scores
          in: body
          schema:
            type: array
            items:
              type: object
              properties:
                studentGithubId:
                  type: string
                mentorGithubId:
                  type: string
                courseTaskId:
                  type: integer
                comment:
                  type: string
                githubPrUrl:
                  type: string
                score:
                  type: integer
      produces:
        - application/json
      responses:
        '200':
          description: ''
  '/course/{courseId}/score':
    post:
      description: Save course task score
      parameters:
        - name: courseId
          in: path
          required: true
          type: integer
      produces:
        - application/json
      responses:
        '200':
          description: ''
    get:
      description: Get course score data
      parameters:
        - name: courseId
          in: path
          required: true
          type: integer
      produces:
        - application/json
      responses:
        '200':
          description: ''
  '/course/{courseId}/expulsion':
    post:
      description: Expel student from course by active mentor
      security:
        - cookieAuth: []
      parameters:
        - name: courseId
          in: path
          description: Course Id
          required: true
          type: integer
        - name: data
          in: body
          schema:
            type: object
            properties:
              studentId:
                type: string
              comment:
                type: string
      produces:
        - application/json
      responses:
        '200':
          description: result
  '/course/{courseId}/feedback':
    post:
      description: Post feedback about a mentor or student
      security:
        - cookieAuth: []
      parameters:
        - name: courseId
          in: path
          description: Course Id
          required: true
          type: integer
        - name: data
          in: body
          schema:
            type: object
            properties:
              toUserId:
                type: integer
              text:
                type: string
              badgeId:
                type: string
      produces:
        - application/json
      responses:
        '200':
          description: result
  '/course/{courseId}':
    get:
      description: Return info about course
      parameters:
        - name: courseId
          in: path
          required: true
          type: integer
      produces:
        - application/json
      responses:
        '200':
          description: ''
  /course:
    post:
      description: Create course
      security:
        - cookieAuth: []
      produces:
        - application/json
      responses:
        '200':
          description: ''
  /courses:
    get:
      description: Gets courses info
      produces:
        - application/json
      responses:
        '200':
          description: Courses info
  /profile:
    get:
      description: get student profile
      security:
        - cookieAuth: []
      produces:
        - application/json
      responses:
        '200':
          description: profile
  /session:
    get:
      description: Gets current user session
      produces:
        - application/json
      responses:
        '200':
          description: operation status
  /tasks:
    post:
      description: Add/Update tasks
      security:
        - cookieAuth: []
      produces:
        - application/json
      responses:
        '200':
          description: operation status
    get:
      description: Gets tasks
      security:
        - cookieAuth: []
      produces:
        - application/json
      responses:
        '200':
          description: List of tasks
  /users:
    post:
      description: Update user activist status
      security:
        - cookieAuth: []
      produces:
        - application/json
      responses:
        '200':
          description: operation status
  '/users/search/:searchText':
    post:
      description: Search users
      security:
        - cookieAuth: []
      produces:
        - application/json
      responses:
        '200':
          description: operation status
  /v2/me:
    get:
      description: Returns users profile
      produces:
        - application/json
      responses:
        '200':
          description: User object
definitions: {}
responses: {}
parameters: {}
securityDefinitions: {}
tags: []